Exercício 2 Feito!!!

+++ src/Exercicio2/Carne.java
@@ -3,10 +3,10 @@ package Exercicio2;
 public class Carne extends ProdutoPerecivel {
 	
 	public Carne (){
-		super (" Não Informado", 0, 0);
+		super (" Não Informado", 0, 0.0);
 	}
 	
-	public Carne (String nome, int id, float preco){
+	public Carne (String nome, int id, Double preco){
 		super.nome = nome;
 		super.id = id;
 		super.preco = preco;
