+++ src/Exercicio2/Panela.java
@@ -3,10 +3,10 @@ package Exercicio2;
 public class Panela extends ProdutoNaoPerecivel {
 	
 	public Panela(){
-		super("Não Informado", 0, 0);
+		super("Não Informado", 0, 0.0);
 	}
 
-	public Panela (String nome, int id, float preco){
+	public Panela (String nome, int id, Double preco){
 		super.nome = nome;
 		super.id = id;
 		super.preco = preco;
