+++ src/Exercicio2/ProdutoNaoPerecivel.java
@@ -3,10 +3,10 @@ package Exercicio2;
 public class ProdutoNaoPerecivel extends Produto {
 
 	public ProdutoNaoPerecivel(){
-		super("Não Informado", 0, 0);
+		super("Não Informado", 0, 0.0);
 	}
 	
-	public ProdutoNaoPerecivel(String nome, int id, float preco){
+	public ProdutoNaoPerecivel(String nome, int id, Double preco){
 		super.nome = nome;
 		super.id = id;
 		super.preco = preco;
